Remove three (!!!) in a row for clone Megas
Clone Megas with a certain bootloader fail on upload with code with three (!!!) in a row. This is a patch.
Proper fix is to update the bootloader on the Mega2560 clone with the official bootloader. Instructions below.

"There are three possible reasons:
(1) The code has three exclamation points in a row in the code, and when the serial chip sees that escape sequence on the bus (even in an upload) it triggers the bootloader to go nuts / reset. This is an issue with some clones.
http://forum.arduino.cc/index.php?topic=83079.msg2668923#msg2668923

(2) Other people say you just have to update your drivers with the ones contained in the Arduino dirctory. Are you using the Arduino.org IDE (bad) or Arduino.CC IDE (good)? They forked the ecosystem some time back when the partner in Italy ran off with the patents.
"I fixed the problem I was having by getting the newest build of the Arduino software and reinstalling the driver from the Arduino folder."

(3) http://arduino.cc/en/Hacking/DFUProgramming8U2
You may have to update the bootloader manually. These Chinese boards -- while extremely cheap, so good for potentially hazardous situations (like cable testing), do require more DIY on the lower levels. (Someone argued you could buy 12 clones for the price of 1 Ruggeduino.)"